Split database code into separate modules.
